backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LLM & Agents - Use compatible ranges
langchain>=0.1.0,<0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
sentence-transformers>=2.2.0

# Vector Store
faiss-cpu>=1.7.0
numpy>=1.24.0

# Utilities
tiktoken>=0.5.0